Changed the name of profile page to be easier to reach from the website
