+++ frontend/src/router/routes.ts
@@ -52,9 +52,9 @@ const routes: RouteRecordRaw[] = [
 	},
 
 	{
-		path: '/profile2',
+		path: '/profile',
 		component: () => import('src/views/layouts/Main.vue'),
-		children: [{ path: '', component: () => import('src/views/pages/Profile2.vue') }]
+		children: [{ path: '', component: () => import('src/views/pages/ProfilePage.vue') }]
 	},
 
 	{
